requirements.txt
numpy>=1.26,<3
librosa>=0.10.2,<1
soundfile>=0.12.1,<1
audioread>=3.0.1,<4
fastapi>=0.115,<1
uvicorn>=0.30,<1
python-multipart>=0.0.9,<1
packaging>=24,<26
pytest>=8.0,<9
openai>=1.68,<2
python-dotenv>=1.0,<2
setuptools>=65
tensorflow>=2.20,<2.22
tensorflow_hub>=0.16,<0.17
certifi>=2024.0.0
